+++ src/fish_key_reader.cpp
@@ -243,7 +243,7 @@ static void process_input(bool continuous_mode) {
 
 /// Setup our environment (e.g., tty modes), process key strokes, then reset the environment.
 [[noreturn]] static void setup_and_process_keys(bool continuous_mode) {
-    set_interactive_session(session_interactivity_t::implied);
+    set_interactive_session(true);
     set_main_thread();
     setup_fork_guards();
     env_init();